Migrated BelongsToMany to use the new SelectLoader
